% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Set up an experiment}
\arguments{
\item{name}{The name of the experiment}

\item{use, try}{the use and try code blocks}
}
\value{
an object of class \code{Experiment}
}
\description{
Set up an experiment
}
\details{
\strong{Methods}
\describe{
\item{\code{use()}}{
use code block
}
\item{\code{try()}}{
try code block
}
}
}
\examples{
library(scientist)
res <- Experiment$new(name = "jane")
res$use({
  x = 5
  x^2
})
res$try({
  y = 5
  y^2
})
res
res$run()
res$use_result
res$try_result
res$result()
}
\keyword{datasets}
