% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\value{
an object of class \code{Experiment}
}
\description{
Experiment

Experiment
}
\details{
\strong{Methods}
\describe{
\item{\code{control(...)}}{
control code block
}
\item{\code{candidate(...)}}{
candidate code block
}
\item{\code{run(...)}}{
execute code, runs \code{call_block()}, then \code{do_comparison()}
\code{...}: pass on parameters through \code{call_block()} to
\code{\link[callr:r]{callr::r()}} (\code{wait=TRUE}) or \code{\link[callr:r_bg]{callr::r_bg()}} (\code{wait=FALSE})
}
\item{\code{call_block()}}{
execute code, and collecting timing data
}
\item{\code{do_comparison()}}{
mostly an internal fxn, run after \code{call_block}
}
\item{\code{result()}}{
fetch the result (named list)
}
\item{\code{publish(browse = TRUE)}}{
publish results. creates an html file. if \code{browse=TRUE}
the file opens in your default browser. if \code{browse=FALSE}
you get a file path.
}
\item{\code{diff()}}{
compare results by "diffing" in the R console. used inside of publish()
to prepare diffs. different diff representations for different object
types:
\itemize{
\item numeric/integer: take difference between 2, or matrix comparing
> 2 results
\item character/factor: highlight differences in strings, like git diffs
\item data.frame/matrix: use \pkg{visdat}
\item images: use \pkg{vdiffr}
\item list: not sure how to do this
\item environments: not sure how to do this
}
}
\item{\code{compare(x = NULL)}}{
set a custom comparison function, must result in a single
boolean
fun: a function
}
}
}
\section{Terminology}{

The \code{control} block is the control, the current state of the code.
The \code{candidate} block is the candidate, or the new version of the code
you want to compare to the control.
}

\examples{
library(scientist)

# basic eg
res <- Experiment$new(name = "jane")
res$control({
  x = 5
  x^2
})
res$candidate({
  y = 5
  y^3
})
res
res$run()
res$control_result
res$candidate_results
res$result()
# publish results
res$publish()

# many candidates
res <- Experiment$new(name = "doe")
res
res$control(stuff = {
  Sys.sleep(2)
  x = 5
  x^2
})
res$candidate(foo = {
  Sys.sleep(2)
  y = 5
  y^3
}, bar = {
  Sys.sleep(2)
  w = 1000
  (w - 20) / 34
})
res
res$run()
res$control_result
res$candidate_results
res$result()
res$times
# publish results
res$publish()

# raise errors when the control and experiment do no match
res <- Experiment$new(name = "treetest", error_on_mismatch = TRUE)
res$control({
  x = 5
  x^2
})
res$candidate({
  y = 5
  y^3
})
res
\dontrun{res$run()}

# if not waiting, run $status() and $collect()
res <- Experiment$new(name = "junipers", wait = FALSE)
res$control({
  x = 5
  x^2
})
res$candidate({
  y = 5
  y^3
})
\dontrun{res$status()}
res$run()
res
res$status()
\dontrun{res$collect()}
res$result()

# set explicit comparison
# FIXME: not working yet
# res <- Experiment$new(name = "jane")
# res$control({
#   x = 5
#   x^2
# })
# res$candidate({
#   y = 5
#   y^2
# })
# res$compare(function(control, candidate) {
#   control/2 == candidate/1
# })
# res$run()
# res$result()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-print}{\code{Experiment$print()}}
\item \href{#method-call_block}{\code{Experiment$call_block()}}
\item \href{#method-status}{\code{Experiment$status()}}
\item \href{#method-collect}{\code{Experiment$collect()}}
\item \href{#method-control}{\code{Experiment$control()}}
\item \href{#method-candidate}{\code{Experiment$candidate()}}
\item \href{#method-run}{\code{Experiment$run()}}
\item \href{#method-compare}{\code{Experiment$compare()}}
\item \href{#method-do_comparison}{\code{Experiment$do_comparison()}}
\item \href{#method-result}{\code{Experiment$result()}}
\item \href{#method-publish}{\code{Experiment$publish()}}
\item \href{#method-diff}{\code{Experiment$diff()}}
\item \href{#method-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(name, error_on_mismatch = FALSE, wait = TRUE, progress = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the experiment}

\item{\code{error_on_mismatch}}{(logical) whether to error on mismatch of
results. default: \code{FALSE}}

\item{\code{wait}}{(logical) wait for code to execute. if \code{FALSE}, code is run in the
background, and you have to run \verb{$collect()} to collect results.
default: \code{TRUE}}

\item{\code{progress}}{(logical) whether to turn on progress information or not,
default: \code{TRUE} (IGNORED RIGHT NOW)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call_block"></a>}}
\subsection{Method \code{call_block()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$call_block(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status"></a>}}
\subsection{Method \code{status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collect"></a>}}
\subsection{Method \code{collect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$collect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control"></a>}}
\subsection{Method \code{control()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$control(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candidate"></a>}}
\subsection{Method \code{candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$candidate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$run(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare"></a>}}
\subsection{Method \code{compare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$compare(x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_comparison"></a>}}
\subsection{Method \code{do_comparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$do_comparison()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-result"></a>}}
\subsection{Method \code{result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$result()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-publish"></a>}}
\subsection{Method \code{publish()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$publish(browse = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diff"></a>}}
\subsection{Method \code{diff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$diff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
